feat: moved limited_s3_user to util
