+++ python/pyarrow/tests/conftest.py
@@ -314,22 +314,3 @@ def s3_server(s3_connection):
             if proc is not None:
                 proc.kill()
 
-
-@pytest.fixture(scope='session')
-def limited_s3_user(request, s3_server):
-    def _limited_s3_user(policy):
-        if sys.platform == 'win32':
-            # Can't rely on FileNotFound check because
-            # there is sometimes an mc command on Windows
-            # which is unrelated to the minio mc
-            pytest.skip('The mc command is not installed on Windows')
-        request.config.pyarrow.requires('s3')
-        tempdir = s3_server['tempdir']
-        host, port, access_key, secret_key = s3_server['connection']
-        address = '{}:{}'.format(host, port)
-        if not _configure_limited_user(tempdir, address,
-                                       access_key, secret_key, policy):
-            pytest.skip(
-                'Could not locate mc command to configure limited user')
-
-    return _limited_s3_user
