+++ python/pyarrow/tests/test_dataset.py
@@ -35,7 +35,8 @@ import pyarrow.csv
 import pyarrow.feather
 import pyarrow.fs as fs
 from pyarrow.tests.util import (change_cwd, _filesystem_uri,
-                                FSProtocolClass, ProxyHandler)
+                                FSProtocolClass, ProxyHandler,
+                                limited_s3_user)
 
 try:
     import pandas as pd
@@ -4354,7 +4355,7 @@ _minio_put_only_policy = """{
 
 @pytest.mark.parquet
 @pytest.mark.s3
-def test_write_dataset_s3_put_only(s3_server, limited_s3_user):
+def test_write_dataset_s3_put_only(s3_server):
     from pyarrow.fs import S3FileSystem
 
     # write dataset with s3 filesystem
